--- next.config.js
@@ -1,11 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-    output: 'export',
-    basePath: "",
-    images: {
-        unoptimized: true,
-    },
-    reactStrictMode: true,
-}
-
-module.exports = nextConfig
