+++ next.config.mjs
@@ -1,0 +1,22 @@
+import remarkGfm from 'remark-gfm'
+import createMDX from '@next/mdx'
+
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+    output: 'export',
+    basePath: "",
+    pageExtensions: ['js', 'jsx', 'mdx', 'ts', 'tsx'],
+    images: {
+        unoptimized: true,
+    },
+    reactStrictMode: true,
+}
+
+const withMDX = createMDX({
+    // Add markdown plugins here, as desired
+    options: {
+        remarkPlugins: [remarkGfm],
+    },
+})
+
+export default withMDX(nextConfig)
